Add api/rides route and refactor for clarity
I moved a lot of stuff around here, primarily taking the Java Spark setup into a method called configureSpark(), which is called right away in main. I did this to put more emphasis on the routes in main rather than the peripheral setup. Open to more refactoring here if need be!