Refresh ride page upon add-ride, remove home and user page
Fixes bug #9 whereupon the ride listing didn't show a new ride on the
ride list page when it was added until the page would be refreshed.
Add-ride component now uses the refreshRide method upon adding a new
ride and that resolves our issue.

We also removed home and user page from the navbar. In addition, the
default localhost:9000 is the rides page.

Close #9